defun: in which I start to become enlightened
I'm pretty sure every lisp geek has a story about when they became enlightened. This is mine.

So here's the interesting thing about lisp enlightenment: you always think it's already happened. In my case, I've been studying and playing with various lisps off and on for years now. I understand code-as-data and I understand how homoiconicity helps that along. When I say I understand it I mean that I know about it and I can tell you why it's important and the kind of things it lets you do. But today was the first day that I *felt* it.

What did it take to become enlightened? Struggling with errors. I was trying to add "defun" to the prelude. It kept not working. I really wished I had a debugger so I could see what lisp was doing. But then I realized: I DO have a debugger. The REPL is right there! So I just kept building up a list in the repl using the list and quote operators and hitting the up arrow and editing it until I had about what I wanted. Then, when I thought it looked good, I "evau"d it. I was able to build up code incrementally, programmatically, until it took the form I wanted. Then I was able to build a function that spit out that code. Then I wrapped that in a define. And then I was done.

Am I fully enlightened? No. But I'm really starting to get it.