(:: define ::)
This introduces the "::" operator, which should be read "as" or "is now". This unifies some previously ad-hoc syntax. For example, now let can be a binding while "let::" says "this is NOW a binding (that will persist)". Now we don't need set.

This is most clear in the "fn" case. "fn" IS a function, "fn::" says "this is NOW a function (with a name that will persist)".